Game Over Screen
Made Game Over Screen and Script. The script is on the Podium Game Object, and you only need to change the names in the inspector or on the public var in order to get the podium working. If you leave a slot blank that slot on the podium will have no tank and no name